ARM OMAP2+ PM: Centralize PM Initialization
Currently different PM layers  are initialized in different
places and there is no specific order or order depends on the
link order in makefile or it will depend on __initcall. This is an attempt
to initialize all PM layers from a single place so that dependencies are
properly taken care. Now PM init is done as a late init call so that all
dependencies are initialized before PM get initialized.
If driver tries to use PM APIs before PM Init, it would fail. So drivers
are expected to wait till PM init is done before PM APIs can be used.

Also omap cpufreq driver initialization is moved to lateinit_sync stage
so that omap cpufreq is initialized only after OMAP PM initialization.

Change-Id: Ib524428496538e2f41cd7e01445be0932a1f5b9a
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>
[grygorii.strashko@ti.com: rebased, move PM initialization on
subsys_initcall_sync layer to allow PMIC be initialized before]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>

Conflicts:

	arch/arm/mach-omap2/pm.c