mmc: core: Allow PowerOff Notify only for eMMC v4.5
For all e.MMC cards that are not following JEDEC v4.51
standard PowerOff Notification should be restricted.
For devices that are comply v4.41 standard suspend/resume
sequence should be performed via CMD5.

So, restricting sending of Power-Off Notify only for devices
that are supported this feature and thereby fixing suspend/resume
for v4.41 devices.

Change-Id: I48e2ce921004447804ef652da11f36245abdb799
Signed-off-by: Riazantsev Volodymyr <v.riazantsev@ti.com>