mmc: block: Fix suspend/resume for mmc_block devices
Assumption that is:

 /*
  * Resume involves the card going into idle
  * so current partition is always the main one.
  */

Is not correct in case if card is support Sleep-state
and this state is used during device suspend/resume
sequence. In Sleep-state card keep all its settings and
thereby current partition as well.

For prevent mmc_block layer from misalignment after resume
we need to keep current partition same that we had before
suspend.

Change-Id: I78dcc0e603e63964bb2313f67a0dfe6f01bc86c2
Signed-off-by: Riazantsev Volodymyr <v.riazantsev@ti.com>