Input: gpio_keys: Add HACK to fix suspend/resume press event lost
During suspend/resume the key press can be lost if time of resume
sequence is significant.

For board we can lose press HOME key event, as a result we resume
then suspend right away because the key state is not changed. So we
need some workaround to emulate key press then key up events.

This patch adds emulation if key press is lost during resume sequence.
Now, if key press event is lost we generate it then in 20ms generate
up event.

It is uncommon to change the general framework which is why it is
marked as HACK.

Change-Id: I4235fa4d4e5addb517c140d0dede50a7a0f8bfcd
Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@ti.com>