twl603x_battery: Fix race in probe
In probe was implemented such sequence:
1. request irqs (twl6030charger_fault_interrupt,
   twl6030charger_ctrl_interrupt)
2. init works (which are called from irq)
3. unmask irqs

The race occurs because USB driver unmasks
TWL6030_CHARGER_CTRL_INT before it must be
unmasked in twl6030_bci_battery_probe. Thus
at some conditions (e.g. on TWL6030 if battery
emulator is enabled) irq starts to handle
after point 1 and before point 2 is reached.
The irq tries to schedule uninited work. This
causes kernel instability or a deadlock.

In this patch the request of the irqs are
moved to the end of the twl6030_bci_battery_probe.
This helps to avoid races and is safe if
irqs are unmasked before probe ends and the
battery isn't ready to charge.

Change-Id: Icb9d35a20933bafdd23d9c06e63d6252db3087e3
Signed-off-by: Oleg_Kosheliev <oleg.kosheliev@ti.com>